#50 refactor: stage-view 専用マーカーを _components 配下へ移動
- current-position-marker / planned-path-marker は stage-view からのみ使用
- 親と同列だった構成を stage-view/_components 配下へ集約
