+++ src/prototypes/time-control/time-control-03/_components/stage-view/_components/planned-path-marker/index.tsx
@@ -1,9 +1,9 @@
 import { memo } from 'react'
 
-import { useStageTransform } from '../../_contexts/stage-transform-context'
-import { toPixelStyle } from '../../_lib/stage-coords'
-import { usePlannedPathStore } from '../../_stores'
-import { ActorId } from '../../types'
+import { useStageTransform } from '../../../../_contexts/stage-transform-context'
+import { toPixelStyle } from '../../../../_lib/stage-coords'
+import { usePlannedPathStore } from '../../../../_stores'
+import { ActorId } from '../../../../types'
 
 type PlannedPathMarkerProps = {
   /** 表示対象の actor */
